facial_recognition_app/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
streamlit>=1.28.0
numpy>=1.24.0,<2.0.0
pillow>=9.5.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
mtcnn>=0.1.1
facenet-pytorch>=2.5.3
tqdm>=4.65.0
pandas>=2.0.0
seaborn>=0.12.0
albumentations>=1.3.0
tensorflow>=2.12.0
